requirements.txt
albumentations==1.4.14
kornia==0.7.3
MedPy==0.5.2
numpy==2.1.3
pandas==2.2.3
Pillow==11.0.0
PyYAML==6.0.2
skimage==0.0
tensorboardX==2.6.2.2
tifffile==2024.7.2
torch==2.1.2+cu118
tqdm==4.64.1
